minor refactor lwc page
