+++ doradb-storage/src/lib.rs
@@ -11,6 +11,7 @@ pub mod free_list;
 pub mod index;
 pub mod latch;
 pub mod lifetime;
+pub mod lwc;
 pub mod memcmp;
 pub mod notify;
 pub mod ptr;
